% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_device_data.R
\name{fetch_device_data}
\alias{fetch_device_data}
\title{Fetch data from a given device}
\usage{
fetch_device_data(mac_address, date = NULL)
}
\arguments{
\item{mac_address}{Device MAC address}

\item{date}{Specific date to retrieve. If blank, device data from the last 24 hours is returned}
}
\description{
Fetch data from a given device
}
