% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_set_api_key.R
\name{get_or_set_api_key}
\alias{get_or_set_api_key}
\title{Get or set Ambient Weather API key}
\source{
Environment variable approach (and code) based on the \href{https://github.com/hrbrmstr/darksky}{darksky} package by Bob Rudis
}
\usage{
get_or_set_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting new API key in current environment}
}
\description{
Get or set Ambient Weather API key
}
