% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_devices.R
\name{list_user_devices}
\alias{list_user_devices}
\title{List user's available devices and each device's most recent data}
\usage{
list_user_devices()
}
\description{
List user's available devices and each device's most recent data
}
